re-estruturando para melhorar legibilidade
