+++ gpon/portas/CTO.py
@@ -1,0 +1,12 @@
+class CTO:
+
+    def __init__(self, l, e, c, df=0, ds=0, o=0, r=0, v=0, t=0):
+        self.localidade = l
+        self.estacao    = e
+        self.cto        = c
+        self.defeito    = df
+        self.designado  = ds
+        self.ocupado    = o
+        self.reservado  = r
+        self.vago       = v
+        self.total      = t
